* commit '0f43513e4ed47953b5e1ba3f42fab9b6f4ea1406 [formerly 42f85e8ae56e1c88ed963c6f92696e4a0c15f7eb]': (249 commits)
  [DF-185] Android compile fix
  DF-2870 define refactoring
  [DF-12672] Code format only
  DF-12798 fix deploy win10 cmake
  DF-11619 texture size cap for metal-api
  DF-11619 texture size cap for DX9, DX11, GL
  DF-12801 Fix path for Android
  [DF-12760] Added missing namespace
  [DF-12760] Fix for missing CoreV2 functionality
  DF-12402 fix after merge
  [DF-12793] We should check current focused widget if context == WidgetContext.
  DF-12166: fixed lack of AnyCasts folderin Android.mk
  DF-12175 try fix android build win ndk-build
  [DF-12760] Fixed bad copypaste
  [DF-185] Generate solution fix
  [DF_12755] Removed wrong space in string
  DF-12175 rename locals
  DF-12175 rename packManager to dlcManager
  DF-12175 rename folder more
  DF-12175 rename folder
  ...


Former-commit-id: e12412c3cf74644c5908984742978062bb19154b

+++ Sources/Internal/Render/2D/TextBlock.h
@@ -107,7 +107,7 @@ public:
     TextBlock* Clone();
     void CopyDataFrom(TextBlock* block);
 
-    const Vector<int32>& GetStringSizes();
+    const Vector<float32>& GetStringSizes();
         
 
 #if defined(LOCALIZATION_DEBUG)
@@ -192,7 +192,7 @@ private:
     WideString logicalText;
     WideString visualText;
     Vector<WideString> multilineStrings;
-    Vector<int32> stringSizes;
+    Vector<float32> stringSizes;
 
     bool isMultilineEnabled : 1;
     bool isMultilineBySymbolEnabled : 1;
